requirements.txt
langchain
langchain-ollama
streamlit
python-dotenv